Changes for 64 bit platforms
git-svn-id: http://svn.cern.ch/guest/lhcb/Online/trunk@96545 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: cfca1727330fad3370dd7361e0caef0c115d0bd6

+++ Online/Online/GaudiOnline/src/MEPConverterSvc.cpp
@@ -153,9 +153,9 @@ int LHCb::MEPConverterSvc::declareSubEvent(const EventDesc& evt, int evID, const
     h->setOffsetOfMissing(0);
     for(size_t j=0; j<frags.size(); ++j)  {
       LHCb::MEPFragment* f = frags[j];
-      int off =  int(int(f)-id->mepStart);
-      // printf("mep:%p Event:%d fragment:%p %d offset:%d size:%d\n",
-      //        id->mepStart,evID,f,j,off,int(f->size()));
+      long off =  long(long(f)-id->mepStart);
+      // printf("mep:%p Event:%d fragment:%p %d offset:%ld size:%ld\n",
+      //        id->mepStart,evID,f,j,off,long(f->size()));
       h->setOffset(j, off);
     }
     e.mask[0] = id->partitionID;
